move parition value realted tests to correct file
